Rag_competition_advanced/requirements.txt
pypdf
langchain
langchain-openai
langchain-community
faiss-cpu
python-dotenv
